# [GSOC] Speeding-up AKAZE, part #2