… modules/features2d/src/kaze/AKAZEFeatures.h
… modules/features2d/src/kaze/TEvolution.h
… modules/features2d/src/kaze/nldiffusion_functions.cpp
… modules/features2d/src/kaze/nldiffusion_functions.h
… modules/features2d/src/opencl/akaze.cl
… modules/features2d/test/test_akaze.cpp
… modules/features2d/test/test_descriptors_invariance.cpp
… modules/features2d/test/test_detectors_regression.cpp
… modules/imgproc/src/smooth.cpp