This part focuses on performance improvements for AKAZE on CPU and implementing basic OCL support.

cc: @bmagyar

OpenCL:
=======

some parts (mainly construction of the pyramid) execute OpenCL paths.

* `Create_Nonlinear_Scale_Space` almost all operatiins are OpenCl enabled
    * GaussianBlur - executes OCL, but less effective path through sepFilter2D
    * Scharr - has ocl
    * resize - has ocl
    * kfactor estimation - needs custom ocl kernel, might reuse some parts for computing histogram
    * diffussion + conductivity - needs custom ocl kernel
    * `Compute_Determinant_Hessian_Response` executes OpenCL general sepFilter2D, computing determinant would need custom kernel, but is not currently blocker
* `Feature_Detection` no OCL
* `Compute_Keypoints_Orientation` no OCL
* `Compute_Descriptors` no OCL


CPU status:
===========

* `Create_Nonlinear_Scale_Space` reworked, some intrinsics might help with diffusion
    * `Compute_Determinant_Hessian_Response` reworked, needs specialized fine tuning for kernels 5x5, 7x7, 9x9
* `Feature_Detection` this will be reworked together with GPU part, we might want the same format of keypoints on GPU and CPU
* `Compute_Keypoints_Orientation` reworked
* `Compute_Descriptors` not a blocker

The main parts (and largest bottlenecks) are:
=============================================

* `Create_Nonlinear_Scale_Space` ~19% ~21ms
    * needs to be reworked
* `Feature_Detection` ~52% ~60ms, of which:
    * `Compute_Determinant_Hessian_Response` 83% (over 43% of the whole algorithm)
        * reworked slightly -> reduced Feature_Detection to ~58ms
    * `Find_Scale_Space_Extrema` 16% (globally just 8%)
        * this might be worth to look into when other parts will be optimized
    * `Do_Subpixel_Refinement` 1%
        * not a bottleneck in the current state
* `Compute_Keypoints_Orientation` ~21% ~24ms
     * reworked completely, replaced by a faster implementation, parallelized -> reduced to 3.69ms
* `Compute_Descriptors` ~4% ~4.8ms
    * not a bottleneck in the current state

Improvement in `Compute_Determinant_Hessian_Response` is of cause not very
satisfying. The main problem there is that AKAZE uses Scharr with non-standard
size kernel (different from 3x3). Implementing Scharr with sepFilter2D is much
slower than specialized `cv::Scharr` we have.

I have tried to replace the `sepFilter2D` implementation with `cv:Scharr`,
just to get the idea of possible speedup. It is in the separate [branch](https://github.com/hrnr/opencv/compare/akaze_part2...hrnr:test_scharr). With
`cv::Scharr` `Compute_Determinant_Hessian_Response` went from ~47ms to ~11ms,
which is quite significant. However it is not possible to replace Scharr just
like, so the tests are failing. I have opened pablofdezalc/akaze#32 to get
some addition info on this.

Performance stats per commit:
=============================

I'm testing all CPU performance on `Intel(R) Core(TM) i7-6700 CPU @ 3.40GHz`.

```

Geometric mean

                                                   Name of Test                                                        perf       perf       perf       perf       perf       perf       perf       perf       perf       perf       perf       perf       perf       perf       perf       perf       perf       perf       perf
                                                                                                                    8200996b1  b12facf48  aa5a72b46  8cc0b286c  1d3f7fe9e  c13351891  76151e566  ea089a8ab  f9c2951fa  ba071d1ad  b12facf48  aa5a72b46  8cc0b286c  1d3f7fe9e  c13351891  76151e566  ea089a8ab  f9c2951fa  ba071d1ad
                                                                                                                                                                                                                                      vs         vs         vs         vs         vs         vs         vs         vs         vs
                                                                                                                                                                                                                                     perf       perf       perf       perf       perf       perf       perf       perf       perf
                                                                                                                                                                                                                                  8200996b1  8200996b1  8200996b1  8200996b1  8200996b1  8200996b1  8200996b1  8200996b1  8200996b1
                                                                                                                                                                                                                                  (x-factor) (x-factor) (x-factor) (x-factor) (x-factor) (x-factor) (x-factor) (x-factor) (x-factor)
detectAndExtract::feature2d::(AKAZE_DEFAULT, "cv/detectors_descriptors_evaluation/images_datasets/leuven/img1.png") 71.572 ms  67.583 ms  44.742 ms  45.349 ms  42.336 ms  40.726 ms  38.732 ms  38.268 ms  43.283 ms  38.549 ms     1.06       1.60       1.58       1.69       1.76       1.85       1.87       1.65       1.86
detectAndExtract::feature2d::(AKAZE_DEFAULT, "stitching/a3.png")                                                    56.269 ms  52.597 ms  34.432 ms  34.941 ms  32.298 ms  31.299 ms  29.619 ms  29.137 ms  31.197 ms  29.492 ms     1.07       1.63       1.61       1.74       1.80       1.90       1.93       1.80       1.91
detectAndExtract::feature2d::(AKAZE_DEFAULT, "stitching/s2.jpg")                                                    273.459 ms 263.874 ms 163.169 ms 168.078 ms 153.912 ms 164.177 ms 147.598 ms 145.255 ms 156.495 ms 150.994 ms    1.04       1.68       1.63       1.78       1.67       1.85       1.88       1.75       1.81

```

If you want to reproduce these results you on your own machine, you might find this [performance evaluation script](https://gist.github.com/hrnr/045b5e8efdfd4713957ce0be2ba2f23a) helpful. It automates running perf tests for this pull request across many revisions, it generates this summary along with xml test outputs and instrumentation output as shown below.

Instrumentation output:
=======================

INITIAL:
--------

```
Time compensation is 0
CTEST_FULL_OUTPUT
OpenCV version: 3.2.0-dev
OpenCV VCS version: 3.2.0-816-g28d66b332
Build type: release
Parallel framework: tbb
CPU features: popcnt mmx sse sse2 sse3 ssse3 sse4.1 sse4.2 avx avx2 fma3 fp16
cl_get_gt_device(): error, unknown device: ffffffff
cl_get_gt_device(): error, unknown device: ffffffff
cl_get_gt_device(): error, unknown device: ffffffff
OpenCL is disabled
Note: Google Test filter = feature2d_detectAndExtract.detectAndExtract/4
[==========] Running 1 test from 1 test case.
[----------] Global test environment set-up.
[----------] 1 test from feature2d_detectAndExtract
[ RUN      ] feature2d_detectAndExtract.detectAndExtract/4
[ PERFSTAT ]    (samples = 13, mean = 114.95, median = 115.11, stddev = 1.16 (1.0%))
[ VALUE    ]    (AKAZE_DEFAULT, "cv/detectors_descriptors_evaluation/images_datasets/leuven/img1.png")
[ TRACE    ]
ROOT
\---detectAndCompute - TC:1 C:13 T:1513.67ms
    |---convertTo - TC:1 C:13 T:2.22ms L:0% G:0%
    |   \---ipp_convertTo<W_IPP> - TC:1 C:13 T:2.18ms L:98% G:0%
    |       \---::ipp::iwiScale<F_IPP> - TC:1 C:13 T:2.17ms L:99% G:0%
    |
    |---zeros - TC:1 C:1664 T:0.75ms L:0% G:0%
    |---operator= - TC:1 C:1664 T:55.10ms L:4% G:4%
    |---Create_Nonlinear_Scale_Space - TC:1 C:13 T:281.87ms L:19% G:19%
    |   |---copyTo - TC:1 C:182 T:17.01ms L:6% G:1%
    |   |   \---ippicviCopy_8u_C1R_L<F_IPP> - TC:1 C:182 T:16.84ms L:99% G:1%
    |   |
    |   |---GaussianBlur - TC:1 C:221 T:50.43ms L:18% G:3%
    |   |   |---ipp_GaussianBlur<W_IPP> - TC:1 C:221 T:27.69ms L:55% G:2%
    |   |   |   |---parallel_for_ - TC:1 C:117 T:25.80ms L:93% G:2%
    |   |   |   |   \---operator() - TC:5 C:2912 T:90.64ms L:351% G:6%
    |   |   |   |       \---operator()<W_IPP> - TC:8 C:2912 T:73.70ms L:81% G:5%
    |   |   |   |           \---::ipp::iwiFilterGaussian<F_IPP> - TC:8 C:2593 T:60.24ms L:82% G:4%
    |   |   |   |               \---::ipp::iwiFilterGaussian - BadExit<MARK_IPP> - TC:8 C:97 T:0.05ms L:0% G:0%
    |   |   |   |
    |   |   |   \---::ipp::iwiFilterGaussian<F_IPP> - TC:1 C:104 T:1.61ms L:6% G:0%
    |   |   |
    |   |   \---sepFilter2D - TC:1 C:13 T:22.39ms L:44% G:1%
    |   |       |---convertTo - TC:1 C:26 T:0.09ms L:0% G:0%
    |   |       |   \---ipp_convertTo<W_IPP> - TC:1 C:26 T:0.06ms L:66% G:0%
    |   |       |       \---::ipp::iwiScale<F_IPP> - TC:1 C:26 T:0.03ms L:60% G:0%
    |   |       |
    |   |       \---apply - TC:1 C:13 T:21.93ms L:98% G:1%
    |   |           |---borderInterpolate - TC:1 C:23608 T:2.98ms L:14% G:0%
    |   |           \---ippiOperator<W_IPP> - TC:1 C:7800 T:8.71ms L:40% G:1%
    |   |               \---ippicviFilterRowBorderPipeline_32f_C1R<F_IPP> - TC:1 C:7800 T:6.43ms L:74% G:0%
    |   |
    |   |---zeros - TC:1 C:143 T:0.09ms L:0% G:0%
    |   |---operator= - TC:1 C:143 T:6.98ms L:2% G:0%
    |   |---Scharr - TC:1 C:416 T:31.74ms L:11% G:2%
    |   |   \---ipp_Deriv<W_IPP> - TC:1 C:416 T:31.45ms L:99% G:2%
    |   |       \---::ipp::iwiFilterScharr<F_IPP> - TC:1 C:416 T:31.04ms L:99% G:2%
    |   |
    |   |---parallel_for_ - TC:1 C:2158 T:72.07ms L:26% G:5%
    |   |   \---operator() - TC:5 C:1612 T:143.77ms L:199% G:9%
    |   |
    |   |---add - TC:1 C:2158 T:29.22ms L:10% G:2%
    |   |   \---ippicviAdd_32f_C1R<F_IPP> - TC:1 C:2158 T:27.56ms L:94% G:2%
    |   |
    |   \---resize - TC:1 C:39 T:2.46ms L:1% G:0%
    |       \---resize - TC:1 C:39 T:2.40ms L:97% G:0%
    |           |---ipp_resize<W_IPP> - TC:1 C:39 T:0.01ms L:1% G:0%
    |           |---parallel_for_ - TC:1 C:26 T:1.56ms L:65% G:0%
    |           |   \---operator() - TC:1 C:26 T:1.49ms L:96% G:0%
    |           |
    |           \---parallel_for_ - TC:1 C:13 T:0.65ms L:27% G:0%
    |
    |---Feature_Detection - TC:1 C:13 T:785.74ms L:52% G:52%
    |   |---Compute_Determinant_Hessian_Response - TC:1 C:13 T:652.72ms L:83% G:43%
    |   |   \---parallel_for_ - TC:1 C:13 T:625.91ms L:96% G:41%
    |   |       \---operator() - TC:1 C:208 T:4019.96ms L:642% G:266%
    |   |           |---copyTo - TC:8 C:2080 T:3.92ms L:0% G:0%
    |   |           |   \---ippicviCopy_8u_C1R_L<F_IPP> - TC:8 C:2080 T:1.48ms L:38% G:0%
    |   |           |
    |   |           |---sepFilter2D - TC:8 C:1040 T:1295.68ms L:32% G:86%
    |   |           |   |---convertTo - TC:8 C:2080 T:6.56ms L:1% G:0%
    |   |           |   |   \---ipp_convertTo<W_IPP> - TC:8 C:2080 T:4.00ms L:61% G:0%
    |   |           |   |       \---::ipp::iwiScale<F_IPP> - TC:8 C:2080 T:1.41ms L:35% G:0%
    |   |           |   |
    |   |           |   \---apply - TC:8 C:1040 T:1287.03ms L:99% G:85%
    |   |           |       |---borderInterpolate - TC:8 C:744900 T:207.46ms L:16% G:14%
    |   |           |       \---ippiOperator<W_IPP> - TC:8 C:219375 T:220.04ms L:17% G:15%
    |   |           |           \---ippicviFilterRowBorderPipeline_32f_C1R<F_IPP> - TC:8 C:219375 T:103.64ms L:47% G:7%
    |   |           |
    |   |           \---convertTo - TC:3 C:1040 T:59.62ms L:1% G:4%
    |   |               \---ipp_convertTo<W_IPP> - TC:3 C:1040 T:56.39ms L:95% G:4%
    |   |                   \---::ipp::iwiScale<F_IPP> - TC:3 C:1040 T:53.24ms L:94% G:4%
    |   |
    |   |---Find_Scale_Space_Extrema - TC:1 C:13 T:122.20ms L:16% G:8%
    |   \---Do_Subpixel_Refinement - TC:1 C:13 T:10.79ms L:1% G:1%
    |       \---solve - TC:1 C:23985 T:5.07ms L:47% G:0%
    |
    |---Compute_Keypoints_Orientation - TC:1 C:13 T:321.29ms L:21% G:21%
    |   \---fastAtan32f - TC:1 C:23946 T:14.92ms L:5% G:1%
    |       \---fastAtan32f - TC:1 C:23946 T:7.91ms L:53% G:1%
    |
    |---Compute_Descriptors - TC:1 C:13 T:63.65ms L:4% G:4%
    |   |---zeros - TC:1 C:13 T:0.01ms L:0% G:0%
    |   |---operator= - TC:1 C:13 T:0.06ms L:0% G:0%
    |   \---parallel_for_ - TC:1 C:13 T:63.48ms L:100% G:4%
    |       \---operator() - TC:8 C:10239 T:149.85ms L:236% G:10%
    |
    \---copyTo - TC:1 C:13 T:0.27ms L:0% G:0%
        \---ippicviCopy_8u_C1R_L<F_IPP> - TC:1 C:13 T:0.24ms L:89% G:0%

IPP weight: 20.2%
OPENCL weight: 0.0%
[/TRACE    ]
[       OK ] feature2d_detectAndExtract.detectAndExtract/4 (1528 ms)
[----------] 1 test from feature2d_detectAndExtract (1528 ms total)

[----------] Global test environment tear-down
[==========] 1 test from 1 test case ran. (1528 ms total)
[  PASSED  ] 1 test.

```

CURRENT (ba071d1ad):
--------------------
```
Time compensation is 0
CTEST_FULL_OUTPUT
OpenCV version: 3.2.0-dev
OpenCV VCS version: 3.2.0-932-g6744b7cc3
Build type: release
Parallel framework: tbb
CPU features: popcnt mmx sse sse2 sse3 ssse3 sse4.1 sse4.2 avx avx2 fma3 fp16
cl_get_gt_device(): error, unknown device: ffffffff
cl_get_gt_device(): error, unknown device: ffffffff
cl_get_gt_device(): error, unknown device: ffffffff
OpenCL is disabled
Note: Google Test filter = feature2d_detectAndExtract.detectAndExtract/6:feature2d_detectAndExtract.detectAndExtract/7:feature2d_detectAndExtract.detectAndExtract/8
[==========] Running 3 tests from 1 test case.
[----------] Global test environment set-up.
[----------] 3 tests from feature2d_detectAndExtract
[ RUN      ] feature2d_detectAndExtract.detectAndExtract/6
[ PERFSTAT ]    (samples = 10, mean = 76.67, median = 76.70, stddev = 1.14 (1.5%))
[ VALUE    ]    (AKAZE_DEFAULT, "cv/detectors_descriptors_evaluation/images_datasets/leuven/img1.png")
[ TRACE    ]
ROOT
\---detectAndCompute - TC:1 C:10 T:766.70ms
    |---convertTo - TC:1 C:10 T:1.90ms L:0% G:0%
    |   \---ipp_convertTo<W_IPP> - TC:1 C:10 T:1.86ms L:98% G:0%
    |       \---::ipp::iwiScale<F_IPP> - TC:1 C:10 T:1.84ms L:99% G:0%
    |
    |---Allocate_Memory_Evolution - TC:1 C:10 T:0.14ms L:0% G:0%
    |---Create_Nonlinear_Scale_Space - TC:1 C:10 T:575.78ms L:75% G:75%
    |   |---GaussianBlur - TC:1 C:170 T:16.65ms L:3% G:2%
    |   |   \---ipp_GaussianBlur<W_IPP> - TC:1 C:170 T:16.28ms L:98% G:2%
    |   |       |---parallel_for_ - TC:1 C:90 T:14.42ms L:89% G:2%
    |   |       |   \---operator() - TC:8 C:2240 T:12.98ms L:90% G:2%
    |   |       |       \---operator()<W_IPP> - TC:8 C:2240 T:12.47ms L:96% G:2%
    |   |       |           \---::ipp::iwiFilterGaussian<F_IPP> - TC:8 C:2240 T:11.99ms L:96% G:2%
    |   |       |
    |   |       \---::ipp::iwiFilterGaussian<F_IPP> - TC:1 C:80 T:1.52ms L:9% G:0%
    |   |
    |   |---copyTo - TC:1 C:10 T:1.63ms L:0% G:0%
    |   |---copyTo - TC:1 C:120 T:9.20ms L:2% G:1%
    |   |   \---ippicviCopy_8u_C1R_L<F_IPP> - TC:1 C:120 T:8.88ms L:97% G:1%
    |   |
    |   |---compute_kcontrast - TC:1 C:10 T:17.30ms L:3% G:2%
    |   |   \---convertTo - TC:1 C:10 T:0.66ms L:4% G:0%
    |   |       \---ipp_convertTo<W_IPP> - TC:1 C:10 T:0.64ms L:97% G:0%
    |   |           \---::ipp::iwiScale<F_IPP> - TC:1 C:10 T:0.63ms L:98% G:0%
    |   |
    |   |---Scharr - TC:1 C:320 T:26.78ms L:5% G:3%
    |   |   \---ipp_Deriv<W_IPP> - TC:1 C:320 T:26.43ms L:99% G:3%
    |   |       \---::ipp::iwiFilterScharr<F_IPP> - TC:1 C:320 T:26.03ms L:98% G:3%
    |   |
    |   |---pm_g2 - TC:1 C:150 T:11.74ms L:2% G:2%
    |   |---parallel_for_ - TC:1 C:1660 T:269.44ms L:47% G:35%
    |   |   \---operator() - TC:8 C:253923 T:205.26ms L:76% G:27%
    |   |       \---nld_step_scalar_one_lane - TC:8 C:253923 T:88.20ms L:43% G:12%
    |   |
    |   |---add - TC:1 C:1660 T:25.94ms L:5% G:3%
    |   |   \---ippicviAdd_32f_C1R<F_IPP> - TC:1 C:1660 T:23.99ms L:92% G:3%
    |   |
    |   |---resize - TC:1 C:30 T:1.99ms L:0% G:0%
    |   |   \---resize - TC:1 C:30 T:1.88ms L:94% G:0%
    |   |       |---ipp_resize<W_IPP> - TC:1 C:30 T:0.01ms L:1% G:0%
    |   |       |---parallel_for_ - TC:1 C:20 T:1.09ms L:58% G:0%
    |   |       |   \---operator() - TC:2 C:20 T:0.80ms L:73% G:0%
    |   |       |
    |   |       \---parallel_for_ - TC:1 C:10 T:0.63ms L:33% G:0%
    |   |
    |   \---Compute_Determinant_Hessian_Response - TC:1 C:10 T:193.26ms L:34% G:25%
    |       \---parallel_for_ - TC:1 C:10 T:193.24ms L:100% G:25%
    |           \---operator() - TC:8 C:160 T:193.08ms L:100% G:25%
    |               |---compute_derivative_kernels - TC:8 C:320 T:1.17ms L:1% G:0%
    |               |   \---copyTo - TC:8 C:640 T:0.72ms L:62% G:0%
    |               |       \---ippicviCopy_8u_C1R_L<F_IPP> - TC:8 C:640 T:0.25ms L:35% G:0%
    |               |
    |               \---sepFilter2D - TC:8 C:800 T:186.08ms L:96% G:24%
    |                   |---convertTo - TC:8 C:1600 T:2.43ms L:1% G:0%
    |                   |   \---ipp_convertTo<W_IPP> - TC:8 C:1600 T:1.42ms L:58% G:0%
    |                   |       \---::ipp::iwiScale<F_IPP> - TC:8 C:1600 T:0.45ms L:32% G:0%
    |                   |
    |                   \---apply - TC:8 C:800 T:183.31ms L:99% G:24%
    |                       \---ippiOperator<W_IPP> - TC:8 C:168750 T:113.84ms L:62% G:15%
    |                           \---ippicviFilterRowBorderPipeline_32f_C1R<F_IPP> - TC:8 C:168750 T:57.25ms L:50% G:7%
    |
    |---Feature_Detection - TC:1 C:10 T:96.31ms L:13% G:13%
    |   |---Find_Scale_Space_Extrema - TC:1 C:10 T:87.95ms L:91% G:11%
    |   \---Do_Subpixel_Refinement - TC:1 C:10 T:8.33ms L:9% G:1%
    |       \---solve - TC:1 C:18400 T:3.88ms L:46% G:1%
    |
    |---Compute_Keypoints_Orientation - TC:1 C:10 T:32.10ms L:4% G:4%
    |   \---parallel_for_ - TC:1 C:10 T:32.10ms L:100% G:4%
    |       \---operator() - TC:8 C:7555 T:30.50ms L:95% G:4%
    |           \---fastAtan2 - TC:8 C:18360 T:18.38ms L:60% G:2%
    |               \---fastAtan32f - TC:8 C:18360 T:11.48ms L:62% G:1%
    |                   \---fastAtan32f - TC:8 C:18360 T:4.18ms L:36% G:1%
    |
    \---Compute_Descriptors - TC:1 C:10 T:54.78ms L:7% G:7%
        \---parallel_for_ - TC:1 C:10 T:54.71ms L:100% G:7%
            \---operator() - TC:8 C:8067 T:54.17ms L:99% G:7%

IPP weight: 17.3%
OPENCL weight: 0.0%
[/TRACE    ]
[       OK ] feature2d_detectAndExtract.detectAndExtract/6 (779 ms)
[ RUN      ] feature2d_detectAndExtract.detectAndExtract/7
[ PERFSTAT ]    (samples = 10, mean = 66.97, median = 67.19, stddev = 0.65 (1.0%))
[ VALUE    ]    (AKAZE_DEFAULT, "stitching/a3.png")
[ TRACE    ]
ROOT
\---detectAndCompute - TC:1 C:10 T:669.65ms
    |---convertTo - TC:1 C:10 T:1.56ms L:0% G:0%
    |   \---ipp_convertTo<W_IPP> - TC:1 C:10 T:1.52ms L:98% G:0%
    |       \---::ipp::iwiScale<F_IPP> - TC:1 C:10 T:1.50ms L:99% G:0%
    |
    |---Allocate_Memory_Evolution - TC:1 C:10 T:0.10ms L:0% G:0%
    |---Create_Nonlinear_Scale_Space - TC:1 C:10 T:539.96ms L:81% G:81%
    |   |---GaussianBlur - TC:1 C:130 T:13.03ms L:2% G:2%
    |   |   \---ipp_GaussianBlur<W_IPP> - TC:1 C:130 T:12.80ms L:98% G:2%
    |   |       |---parallel_for_ - TC:1 C:90 T:11.45ms L:89% G:2%
    |   |       |   \---operator() - TC:8 C:2080 T:10.16ms L:89% G:2%
    |   |       |       \---operator()<W_IPP> - TC:8 C:2080 T:9.68ms L:95% G:1%
    |   |       |           \---::ipp::iwiFilterGaussian<F_IPP> - TC:8 C:2080 T:9.24ms L:95% G:1%
    |   |       |
    |   |       \---::ipp::iwiFilterGaussian<F_IPP> - TC:1 C:40 T:1.10ms L:9% G:0%
    |   |
    |   |---copyTo - TC:1 C:10 T:1.35ms L:0% G:0%
    |   |---copyTo - TC:1 C:90 T:5.47ms L:1% G:1%
    |   |   \---ippicviCopy_8u_C1R_L<F_IPP> - TC:1 C:90 T:5.27ms L:96% G:1%
    |   |
    |   |---compute_kcontrast - TC:1 C:10 T:14.72ms L:3% G:2%
    |   |   \---convertTo - TC:1 C:10 T:0.56ms L:4% G:0%
…