@alalek 2017-06-21
@hrnr Please keep patches in a single PR: #8869

```
# git commit pending changes
git checkout -B akaze_part1 HEAD
git push origin akaze_part1
```