@alalek 2017-06-21
OK, we merged part 1 (#8869). Lets continue here.