@hrnr 2017-06-21
Thanks. Sorry for confusion, I should have add more description, to clarify that the situation.

I will rebase for sure.