@ysolovyov 2017-06-21
How much it speed up the algo on your machine?