@hrnr 2017-06-22
This is work in progress. I will update the description with my measurements. The current improvement is minor.