@hrnr 2017-06-30
rebase due to merge conflict. The commit hashes are now different from what is reported above in the perf stats (I will fix that in the future).