@hrnr 2017-07-07
I have updated the [script](https://gist.github.com/hrnr/045b5e8efdfd4713957ce0be2ba2f23a) for evaluating the performace. It can now execute the relevant OpenCL tests too. Run

```
./evaluate_perf.sh ocl < revisions_ocl.txt
```

for OpenCL summary and

```
./evaluate_perf.sh < revisions.txt
```

for CPU tests (be aware that these can use OpenCL paths too on OpenCL enabled hardware).